Fix handle relatonships when ALLOW_UNKNOWN_INCLUSIONS is true (#124)
* Fix handle relatonships when ALLOW_UNKNOWN_INCLUSIONS is true

The logic in RescourceConverter.parseIncluded() that picks which node
of the included array to pass into handleRelationships can select the
incorrect node when the ALLOW_UNKNOWN_INCLUSIONS flag is true.

The logic assumes that the array of INCLUDED and the array
includeResources are the same size and same order.  This is not
true when the getIncludedResources() method is skipping unknown
resources.

This change uses Maps instead of arrays to explicitly get the correct
node from the INCLUDED objects to pass into handleRelationships

The change also has a update to the testEnableAllowUnknownInclusions()
test and associated unknown-inclusions.json.  The test now checks that
the correct statuses are added to the User object.

* Remove Resource Class and fix formatting

- Remove the Resource Class, since it is no longer used.
- Fix formatting for the if statement
- make sure all indentation is using tabs